Fix SSR crash (WebGL mode)

+++ examples/jsm/tsl/display/SSRNode.js
@@ -1056,9 +1056,9 @@ class SSRNode extends TempNode {
 			// Blur traces a single mirror ray, so spend steps in proportion to the ray's screen-space
 			// length (cheap for the short rays that dominate). Scatter needs a fixed, bounded count for
 			// coherent stochastic sampling; each step then spans the whole ray as rayVec / totalStep.
-			const totalStep = this.stochastic === false
+			const totalStep = int( this.stochastic === false
 				? trunc( max( abs( xLen ), abs( yLen ) ).mul( this.quality.clamp() ) ).max( int( 1 ) ).toConst()
-				: this.quality.clamp().mul( MAX_STEPS ).max( int( 1 ) ).toConst();
+				: this.quality.clamp().mul( MAX_STEPS ).max( float( 1 ) ) ).toConst();
 
 			const xSpan = xLen.div( totalStep ).toVar();
 			const ySpan = yLen.div( totalStep ).toVar();
